Generate WiFi login QR Code
This Python script will create a QR Code to provide a simple method to get your phone or tablet onto a Wi-Fi network.

The QR Code includes the SSID or network name and the password. Scan it and your phone will try and connect to the Wi-Fi network. If the network is close and the password matches your device will connect.